+++ changelog.md
@@ -1,1 +1,8 @@
 # Changelog
+
+### 2020-04-19
+#### [#60b05j](https://app.clickup.com/t/60b05j) - Create class Student in models.py
+- Created a Student class for authorization and Profile class for more information and creating a profile page
+- Added signals for custom user creation and creating a profile automatically
+#### [#60b05r](https://app.clickup.com/t/60b05r) - Create account creation feature
+- Created a form, views and urls for registration and login